+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - added support for /copy-context command
 - added support for /tokens command
 - added token usage progress bar to cost info
+- added direct support for OpenRouter in Agent mode
 
 ## [0.11.1]